+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Fixed
 
+- `spoof-css` scriptlet — incorrect `DOMRect` setting [#498].
 - Escaping quotes in `trusted-replace-node-text` scriptlet [#440].
 - `trusted-suppress-native-method` scriptlet, `isMatchingSuspended` was not reset when the stack does not match,
   so in some cases given method was not prevented [#496].
@@ -44,6 +45,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [#477]: https://github.com/AdguardTeam/Scriptlets/issues/477
 [#496]: https://github.com/AdguardTeam/Scriptlets/issues/496
 [#497]: https://github.com/AdguardTeam/Scriptlets/issues/497
+[#498]: https://github.com/AdguardTeam/Scriptlets/issues/498
 
 ## [v2.1.7] - 2025-04-03